Initial work on getting MCTS to work with a neural network complete.
In the future we may want to change the network architecture naturally, as it is terrible by design, and consider generalizing some things here. Of course, we still need to use all of this stuff in `hash-train`, to be able to actually train our networks, although this should be relatively simple given the work we've done so far. After that the main focus will likely be to optimize the MCTS search and experimenting with different selection policies.